% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_msigdbh.R
\name{recode_msigdbh}
\alias{recode_msigdbh}
\title{Nicely format MSigDB Hallmark gene set names.}
\usage{
recode_msigdbh(msigdbh_names)
}
\arguments{
\item{msigdbh_names}{A vector of MSigDB Hallmark gene set names from MSigDB H GMT files to be formatted.}
}
\value{
A character vector of formatted MSigDB Hallmark gene set names.
}
\description{
American spelling, sentence case, Greek characters included.
}
\examples{
x <- c(
  "HALLMARK_ADIPOGENESIS",
  "HALLMARK_G2M_CHECKPOINT",
  "HALLMARK_IL6_JAK_STAT3_SIGNALING"
)
recode_msigdbh(x)
}
